fix(scripts): agregar casts ::integer faltantes en ajuste de grupo (backfill H5)
El UPDATE del paso 5 (simularAjusteDivisorPorGrupo) reusaba $1 sin cast
explicito contra una columna integer (unidades_programadas) y una numeric
(cantidad_paquetes) en la misma sentencia -- Postgres no puede deducir un
tipo unico para el parametro en esos casos. El codigo real del que se
copio este bloque (masas.controller.js:551-561, sap.controller.js
ajustesGrupoMerge) SI tiene los casts ::integer en cada uso; se me
olvidaron al portarlo al script.

Fallo real contra produccion (543->512, item PANPAQ22 con
multiplo_divisor=20 que si dispara un ajuste de grupo real -- el caso
anterior, 542->500, no lo habia disparado y por eso no se detecto antes):
"inconsistent types deduced for parameter $1". Rollback limpio de la
transaccion del caso, verificado por SQL antes de este fix -- nada quedo
a medias.

+++ backend/scripts/backfill_h5_fusion_adicionales.js
@@ -303,9 +303,9 @@ async function fusionar(client, caso) {
   for (const ajuste of ajustesGrupo) {
     await client.query(
       `UPDATE productos_por_masa
-       SET unidades_programadas = $1, kilos_programados = gramaje_unitario * $1 / 1000.0,
-           cantidad_paquetes = $1, origen_ajuste_divisor = 'MERGE_OV_BACKFILL_H5',
-           unidades_ajuste_grupal = unidades_ajuste_grupal + $2, updated_at = NOW()
+       SET unidades_programadas = $1::integer, kilos_programados = gramaje_unitario * $1::integer / 1000.0,
+           cantidad_paquetes = $1::integer, origen_ajuste_divisor = 'MERGE_OV_BACKFILL_H5',
+           unidades_ajuste_grupal = unidades_ajuste_grupal + $2::integer, updated_at = NOW()
        WHERE id = $3`,
       [ajuste.unidadesProgramadasNuevas, ajuste.deltaPaquetes, ajuste.productoId]
     );
